+++ src/extension.ts
@@ -14,6 +14,7 @@ import { FilterPreviewEditor } from "./preview/FilterPreviewEditor";
 import { CodelensProvider } from "./CodelensProvider";
 import { SoundPlayer } from "./utils/soundPlayer";
 import path from "path";
+import { GameDataService } from "./services/gameDataService";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -190,7 +191,7 @@ export function activate(context: vscode.ExtensionContext) {
   // }
 
   // Register diagnostics
-  registerDiagnostics(context);
+  registerDiagnostics(context, gameData);
 
   // Register code actions
   context.subscriptions.push(